+++ src/app/(table)/layout.js
@@ -1,0 +1,14 @@
+import "../globals.css";
+
+export const metadata = {
+  title: "Game Stats",
+  description: "Get stats from games",
+};
+
+export default function RootLayout({ children }) {
+  return (
+        <>
+            {children}
+        </>
+  );
+}
