+++ src/app/(table)/loading.js
@@ -1,0 +1,7 @@
+const Loading = () => {
+    return (
+        <h1 className="absolute top-1/2 left-1/2">Loading...</h1>
+    );
+}
+
+export default Loading;
