--- src/app/Components/GamesList.js
@@ -1,16 +1,0 @@
-import Game from "./Game";
-
-const GamesList = ({ games }) => {
-
-    const renderedGames = games.map((game) => {
-        return  <Game key={game.name} icon={game.icon} name={game.name} altText={game.altText} />
-    });
-
-    return(
-        <>
-            {renderedGames}
-        </>
-    );
-}
-
-export default GamesList;
