+++ src/app/Components/TableComp/TableInputs.js
