+++ src/app/layout.js
@@ -1,5 +1,6 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
+import Navbar from "./Components/NavbarComp/Navbar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -8,10 +9,14 @@ export const metadata = {
   description: "Get stats from games",
 };
 
+
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <Navbar />
+        {children}
+      </body>
     </html>
   );
 }
